V8_YOLO_OPENCV_DNN/requirements.txt
# YOLOv8
ultralytics>=8.0.0

# OpenCV
opencv-python>=4.8.0

# Numpy
numpy>=1.24.0

# PaddleOCR v5 Mobile
paddlepaddle>=3.2.0
paddleocr>=3.3.3

# Utilities
python-dotenv>=1.0.1
Pillow>=10.0.0
